PERF/TODO: the forty-sixth pass's four rows, three refutations, and the fold
PERF: the pass's Log entry (-1.312 % over four commits, each with its
callgrind pair and `--bench` invariants), a fresh Profile of record at the
46th tip, a `resolve_combat` candidate (-25) — 11.86 % over 2,694, read from
the top for the first time — and a refreshed allocator caller table in (-23)
with two unexplained rows in it. Three things written down as *not* to
retry: the ward gathers (the gate removes a fifth of them), guarding
`declare_blockers`' first cold write (measured: it promotes the next one),
and "`activate_ability_inner` asks `card_type_change_in_scope` twice", which
callgrind says is once.

Also folds PERF back under its own 2.8 k line: the forty-second pass's prose
and the 42nd/44th profile tables become indexes (2,865 -> 2,612).

TODO: NEXT rewritten. The headline for the next run is that the iteration
loop is 4x faster than this file claimed — engine-only `profiling-fast` is
3m15s, not 13 min, once the probe-dir clean stops deleting the catalog.

Suite 18,708 / 0 failed / 5 ignored over 22 binaries, golden traces included;
`cargo clippy --workspace --all-targets` clean (the one warning this pass
introduced, a `usize -> usize` cast in the new `wants_converge` L1 index, is
fixed here).

+++ crabomination_base/src/card.rs
@@ -5298,7 +5298,7 @@ impl CardDefinition {
         }
         let key = self.name.as_ptr() as usize;
         // Key 0 is the empty marker; a `&str`'s data pointer is never null.
-        let slot = (key.wrapping_mul(0x9E37_79B9_7F4A_7C15) >> 48) as usize & (L1 - 1);
+        let slot = ((key as u64).wrapping_mul(0x9E37_79B9_7F4A_7C15) >> 48) as usize & (L1 - 1);
         if let Some(hit) = NEAR.with(|c| {
             let c = c.borrow();
             (c[slot].0 == key).then_some(c[slot].1)
